requirements.txt
Django==2.1.4
pytz==2018.7
djangorestframework
django-oauth-toolkit
django-cors-headers
Pillow
PyMySQL
django-pipeline
six
sorl-thumbnail
tld
lxml
qiniu
redis
channels>=2.0.0
channels_redis
django-celery-results
django-celery-beat
django-redis
celery
git+ssh://git@github.com/swolfod/LushuUtils.git
